align implementation with ios
