+++ packages/expo/android/src/main/java/expo/modules/networkfetch/NativeRequestCredentials.kt
@@ -1,0 +1,10 @@
+// Copyright 2015-present 650 Industries. All rights reserved.
+
+package expo.modules.networkfetch
+
+import expo.modules.kotlin.types.Enumerable
+
+internal enum class NativeRequestCredentials(val value: String) : Enumerable {
+  INCLUDE("include"),
+  OMIT("omit")
+}
