+++ packages/expo/android/src/main/java/expo/modules/networkfetch/NativeRequestInit.kt
@@ -1,0 +1,12 @@
+// Copyright 2015-present 650 Industries. All rights reserved.
+
+package expo.modules.networkfetch
+
+import expo.modules.kotlin.records.Field
+import expo.modules.kotlin.records.Record
+
+internal data class NativeRequestInit(
+  @Field val credentials: NativeRequestCredentials = NativeRequestCredentials.INCLUDE,
+  @Field val headers: List<Pair<String, String>> = emptyList(),
+  @Field val method: String = "GET"
+) : Record
