+++ packages/expo/android/src/main/java/expo/modules/networkfetch/NativeResponse.kt
@@ -3,7 +3,8 @@
 package expo.modules.networkfetch
 
 import android.util.Log
-import expo.modules.kotlin.sharedobjects.SharedRef
+import expo.modules.kotlin.AppContext
+import expo.modules.kotlin.sharedobjects.SharedObject
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -12,11 +13,11 @@ import okhttp3.Callback
 import okhttp3.Response
 import okio.BufferedSource
 import java.io.IOException
-import java.nio.ByteBuffer
 
-internal class NativeResponse(private val coroutineScope: CoroutineScope) :
-  SharedRef<ResponseSink>(ResponseSink()), Callback {
-  var state: ResponseState = ResponseState.INITIALIZED
+internal class NativeResponse(appContext: AppContext, private val coroutineScope: CoroutineScope) :
+  SharedObject(appContext), Callback {
+  val sink = ResponseSink()
+  private var state: ResponseState = ResponseState.INITIALIZED
     get() = synchronized(this) { field }
     private set(value) {
       synchronized(this) {
@@ -34,10 +35,10 @@ internal class NativeResponse(private val coroutineScope: CoroutineScope) :
     private set
 
   val bodyUsed: Boolean
-    get() = this.ref.bodyUsed
+    get() = this.sink.bodyUsed
 
   override fun deallocate() {
-    this.ref.finalize()
+    this.sink.finalize()
     super.deallocate()
   }
 
@@ -47,10 +48,10 @@ internal class NativeResponse(private val coroutineScope: CoroutineScope) :
     }
     if (state == ResponseState.RESPONSE_RECEIVED) {
       state = ResponseState.BODY_STREAMING_STARTED
-      val queuedData = this.ref.finalize()
+      val queuedData = this.sink.finalize()
       emit("didReceiveResponseData", queuedData)
     } else if (state == ResponseState.BODY_COMPLETED) {
-      val queuedData = this.ref.finalize()
+      val queuedData = this.sink.finalize()
       emit("didReceiveResponseData", queuedData)
       emit("didComplete")
     }
@@ -160,7 +161,7 @@ internal class NativeResponse(private val coroutineScope: CoroutineScope) :
         break
       }
       if (state == ResponseState.RESPONSE_RECEIVED) {
-        ref.appendBufferBody(stream.buffer.readByteArray())
+        sink.appendBufferBody(stream.buffer.readByteArray())
       } else if (state == ResponseState.BODY_STREAMING_STARTED) {
         emit("didReceiveResponseData", stream.buffer.readByteArray())
       } else {
@@ -177,45 +178,3 @@ internal class NativeResponse(private val coroutineScope: CoroutineScope) :
 }
 
 private typealias StateChangeListener = (ResponseState) -> Boolean
-
-internal class ResponseSink {
-  private val bodyQueue: MutableList<ByteArray> = mutableListOf()
-  private var isFinalized = false
-  var bodyUsed = false
-    private set
-
-  internal fun appendBufferBody(data: ByteArray) {
-    bodyUsed = true
-    bodyQueue.add(data)
-  }
-
-  fun finalize(): ByteArray {
-    val size = bodyQueue.sumOf { it.size }
-    val byteBuffer = ByteBuffer.allocate(size)
-    for (byteArray in bodyQueue) {
-      byteBuffer.put(byteArray)
-    }
-    bodyQueue.clear()
-    bodyUsed = true
-    isFinalized = true
-    return byteBuffer.array()
-  }
-}
-
-internal enum class ResponseState(val intValue: Int) {
-  INITIALIZED(0),
-  STARTED(1),
-  RESPONSE_RECEIVED(2),
-  BODY_COMPLETED(3),
-  BODY_STREAMING_STARTED(4),
-  BODY_STREAMING_CANCELLED(5),
-  ERROR_RECEIVED(6)
-}
-
-internal data class NativeResponseInit(
-  val headers: List<Pair<String, String>>,
-  val status: Int,
-  val statusText: String,
-  val url: String,
-  val redirected: Boolean
-)
